big file system restructure for easier reading
